+++ jdkjson-core/src/main/java/org/jdkstack/jdkjson/core/path/JsonPathSyntax.java
@@ -1,4 +1,4 @@
-package org.jdkstack.jdkjson.core;
+package org.jdkstack.jdkjson.core.path;
 
 /**
  * .
