+++ jdkjson-core/src/main/java/org/jdkstack/jdkjson/core/reader/verson2/AbstractJsonReaderV2.java
@@ -1,8 +1,9 @@
-package org.jdkstack.jdkjson.core;
+package org.jdkstack.jdkjson.core.reader.verson2;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import org.jdkstack.jdkjson.core.common.Ascii;
 
 /**
  * .
