+++ jdkjson-core/src/main/java/org/jdkstack/jdkjson/core/reader/verson2/JsonReader.java
@@ -1,4 +1,4 @@
-package org.jdkstack.jdkjson.core;
+package org.jdkstack.jdkjson.core.reader.verson2;
 
 /**
  * .
