+++ jdkjson-core/src/main/java/org/jdkstack/jdkjson/core/reader/verson2/JsonReaderV2.java
@@ -1,9 +1,12 @@
-package org.jdkstack.jdkjson.core;
+package org.jdkstack.jdkjson.core.reader.verson2;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
+import org.jdkstack.jdkjson.core.common.Ascii;
+import org.jdkstack.jdkjson.core.common.Constants;
+import org.jdkstack.jdkjson.core.exception.JsonRuntimeException;
 import org.jdkstack.jdkjson.core.cache.LruV1;
 
 /**
